headcannon: adjust filament path, bite, preload
these values were previously gussed, partly based empirical feedback
using the poor idler insert geometry that was not sufficiently rigid.

the new values are based on actual data.

idler preload is set to allow 0.02 mm of play in the MR63 bearing and
0.05 mm of M3 thread radius below 3 mm (expected diameter is 2.927
mm).

bite was measured using actual filament compressed between a pair of
hobs and taking half the value for a single hob.

offset is now computed to have the filament go straight up/down off
the idler bearings.