headcannon: rearrange bolts to eliminate one of the four
the primary goal here is to solve an interference with delta arms from
the bolt point nearest the A tower, but after rotating the worm bolts
the other direction, there's no reason to have one right next to the
worm wheel bolt. instead let one bolt serve both roles, which reduces
mass a bit as a bonus.