ujoint: improve pinch, add variants for motor D shaft and 3 mm worm shaft
the motor shaft variant does not bottom out but lets the shaft go
through, so it can be placed as close as possible to the motor body.

the variant for 3mm shaft has an extra 2 mm of pinch engagement, and
zero clearance, in order to compemnsate for loss of mechanical
advantage in the pinch interface.

the pinch cut for all variants it updated to be longer, so that the
last bit of engagement isn't wasted and unable to pinch tight.