headcannon/noncannonical: improve filament path printability
the old worm wheel side clearance did not take into account that the
hob has an extra 0.9 or more of metal before the worm wheel starts.
this had the cutout cutting unnecessarily low into the filament path
insert, making the bearing holes at least in the noncannonical version
cut through the top to the worm wheel cutout zone. reduce the excess
clearance so that does not happen.

also, the overhang over the bearing holes was sloped a lot more than
it needed to be in order to be printable. this also contributed to the
frailty of the top, so flatten it out a bit. there was already a
factor for this from long ago, but it was never tuned; change it from
1 to 0.7.